requirements.txt
pandas
numpy
faiss-cpu
langchain
langchain-community
sentence-transformers
